+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * Add `resolveType()` for customizing how types are resolved in runtimes extending `SymfonyRuntime`
+ * Add support for apps returning a `Response` object in `FrankenPhpWorkerRunner`
 
 7.4
 ---